backend/requirements.txt
torch
torchvision
opencv-python-headless
fastapi
uvicorn
python-multipart
Pillow
numpy
